When an argument was only a field code, an empty string was sent instead of nothing; fixed.
If an argument was nothing but a field code, we shouldn't send anything, as opposed to send an empty string as argument (which might cause problems, or unexpected behaviors, with some apps. E.g. a file manager or browser would open a new tab in the "current" folder or something...)
Now in such a case we just don't send any argument at all.